pyproject.toml
[project]
name = "crew-try"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "MuhammadRaffey", email = "muhammadraffey26@gmail.com" }]
requires-python = ">=3.11"
dependencies = ["crewai[tools]>=0.102.0", "python-dotenv>=1.0.1"]

[project.scripts]
try = "crew_try.main:kickoff"
plot = "crew_try.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
